Define LoadingViewController and LoadingUI
