+++ BothamUI/BothamViewController.swift
@@ -9,7 +9,6 @@
 import Foundation
 import UIKit
 
-
 public class BothamViewController: UIViewController, BothamUI {
     public var presenter: BothamPresenter! = nil
